Refactor: removing the common parts out of the slow tests into a common module
