+++ slow-tests/performance_tests/common_functions.py
@@ -1,0 +1,28 @@
+import git
+
+
+def shorten_sha(long_sha):
+    return long_sha[:7]
+
+
+def git_current_head_sha():
+    repo = git.Repo(search_parent_directories=True)
+    return repo.head.commit.hexsha
+
+
+def generate_data() -> list:
+    text_with_emojis = "I love ⚽ very much 😁."
+    text_with_a_number = '2833047 people live in this area. It is not a good area.'
+    text_with_two_numbers = '2833047 and 1111 people live in this area.'
+    text_with_punctuations = "This sentence doesn't seem to too many commas, periods or semi-colons (;)."
+    text_with_a_date = "Todays date is 04/28/2020 for format mm/dd/yyyy, not 28/04/2020."
+    text_with_dates = "Todays date is 28/04/2020 and tomorrow's date is 29/04/2020."
+    text_with_duplicates = 'Everyone here is so hardworking. Hardworking people. ' \
+                           'I think hardworking people are a good trait in our company.'
+    data = [text_with_emojis, text_with_a_number, text_with_two_numbers,
+            text_with_punctuations, text_with_a_date, text_with_dates, text_with_duplicates]
+
+    new_data = []
+    for index in range(1):
+        new_data.extend(data)
+    return new_data
